Move dextre into subpackage dnsbl and main to dextre
Rearrange data as preparation for #72 and #73
