+++ mss/dextre/dnsbl/dnsbl.go
@@ -1,4 +1,4 @@
-package dextre
+package dnsbl
 
 import (
 	"fmt"
