+++ mss/dextre/dnsbl/dnsbl_test.go
@@ -1,4 +1,4 @@
-package dextre
+package dnsbl
 
 import "testing"
